cruxintern/requirements.txt
asgiref==3.7.2
attrs==23.2.0
blis==0.7.11
catalogue==1.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
cymem==2.0.8
Django==4.2.10
djangorestframework==3.14.0
docx2txt==0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
idna==3.6
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
pandas==2.2.0
pdfminer.six==20231228
plac==1.1.3
preshed==3.0.9
pycparser==2.21
pycryptodome==3.20.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2024.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sortedcontainers==2.4.0
spacy==2.3.5
sqlparse==0.4.4
srsly==1.0.7
thinc==7.4.6
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wasabi==0.10.1
